+++ src/__tests__/route.test.ts
@@ -1,13 +1,14 @@
 import { AcalaJsonRpcProvider, sleep } from '@acala-network/eth-providers';
 import { CHAIN_ID_BSC, CHAIN_ID_KARURA, hexToUint8Array, parseSequenceFromLogEth, redeemOnEth } from '@certusone/wormhole-sdk';
-import axios from 'axios';
+import axios, { AxiosError } from 'axios';
 import { expect } from 'chai';
 import { ContractReceipt, Wallet } from 'ethers';
 import { JsonRpcProvider } from '@ethersproject/providers';
 import { parseUnits } from 'ethers/lib/utils';
 import { after, before } from 'mocha';
 import { ApiPromise, WsProvider } from '@polkadot/api';
-import { ERC20, ERC20__factory } from '@certusone/wormhole-sdk/lib/cjs/ethers-contracts';
+import { ERC20__factory } from '@certusone/wormhole-sdk/lib/cjs/ethers-contracts';
+
 import {
   ROUTE_XCM_URL,
   SHOULD_ROUTE_XCM_URL,
@@ -200,7 +201,31 @@ describe('/relayAndRoute', () => {
     expect(routerCode).to.eq('0x');
   });
 
-  // describe.skip('when should not route', () => {});
+  it('when should not route (should not relay)', async () => {
+    const routeArgs = {
+      dest,
+      destParaId: BASILISK_PARA_ID,
+      originAddr: ETH_USDC,
+    };
+
+    const relayAndRouteArgs = {
+      ...routeArgs,
+
+      // bridge 0.000001 USDC
+      signedVAA: '010000000001004ba23fa55bcb370773bdba954523ea305f96f814f51ce259fb327b57d985eec86a0f69bf46c4bb444d09b1d70e3b2aaa434639ec3ae93f5d0671b3e38055cf3501648726ae4d36000000040000000000000000000000009dcf9d205c9de35334d646bee44b2d2859712a0900000000000012580f01000000000000000000000000000000000000000000000000000000000000000100000000000000000000000007865c6e87b9f70255377e024ace6630c1eaa37f00020000000000000000000000008341cd8b7bd360461fe3ce01422fe3e24628262f000b0000000000000000000000000000000000000000000000000000000000000000',
+    };
+
+    try {
+      await relayAndRoute(relayAndRouteArgs);
+
+      expect.fail('relayAndRoute did not throw when it should!');
+    } catch (err) {
+      expect((err as AxiosError).response?.data).to.deep.contain({
+        error: 'token amount too small to relay',
+        msg: 'cannot relay this request!',
+      });
+    }
+  });
 });
 
 describe('/routeWormhole', () => {
